requirements.txt
streamlit
langchain
langchain-groq
python-dotenv
SpeechRecognition
gTTS
streamlit-mic-recorder